判断fork工程不存在，退出

+++ filelist.sh
@@ -15,7 +15,8 @@ repo=$(cat ./variables.json |jq -r .name)
 fork_url=$(cat ./variables.json |jq -r .fork_url)
 
 rm -rf /tmp/docker-$repo
-git clone $fork_url /tmp/docker-$repo
+git clone $fork_url /tmp/docker-$repo\
+[[ $? -ne 0 ]] && exit
 rm -rf /tmp/docker-$repo/{.git,.github}
 
 unalias cp 1>/dev/null 2>&1
